Feature/auth middleware

+++ src/user-service/user/default_controller.go
@@ -2,7 +2,6 @@ package user
 
 import (
 	"encoding/json"
-	"hsfl.de/group6/hsfl-master-ai-cloud-engineering/user-service/crypto"
 	"hsfl.de/group6/hsfl-master-ai-cloud-engineering/user-service/user/model"
 	"net/http"
 	"strconv"
@@ -16,20 +15,6 @@ func NewDefaultController(userRepository Repository) *defaultController {
 	return &defaultController{userRepository}
 }
 
-func (controller defaultController) GetUsers(writer http.ResponseWriter, request *http.Request) {
-	values, err := controller.userRepository.FindAll()
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
-	}
-
-	writer.Header().Set("Content-Type", "application/json")
-	err = json.NewEncoder(writer).Encode(values)
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
 func (controller defaultController) GetUsersByRole(writer http.ResponseWriter, request *http.Request) {
 	userRoleStr, ok := request.Context().Value("userRole").(string)
 	if !ok {
@@ -40,6 +25,12 @@ func (controller defaultController) GetUsersByRole(writer http.ResponseWriter, r
 	userRole, err := strconv.ParseUint(userRoleStr, 10, 64)
 	if err != nil {
 		http.Error(writer, "Invalid userRole value", http.StatusBadRequest)
+		return
+	}
+
+	_, exists := request.Context().Value("auth_userId").(int)
+	if !exists {
+		writer.WriteHeader(http.StatusUnauthorized)
 		return
 	}
 
@@ -56,8 +47,17 @@ func (controller defaultController) GetUsersByRole(writer http.ResponseWriter, r
 		return
 	}
 
+	formattedResponse := make([]*JsonFormatGetUserByRoleResponse, len(values))
+	for i, value := range values {
+		formattedResponse[i] = &JsonFormatGetUserByRoleResponse{
+			ID:   value.Id,
+			Name: value.Name,
+			Role: value.Role,
+		}
+	}
+
 	writer.Header().Set("Content-Type", "application/json")
-	err = json.NewEncoder(writer).Encode(values)
+	err = json.NewEncoder(writer).Encode(formattedResponse)
 	if err != nil {
 		http.Error(writer, err.Error(), http.StatusInternalServerError)
 		return
@@ -66,24 +66,33 @@ func (controller defaultController) GetUsersByRole(writer http.ResponseWriter, r
 
 func (controller defaultController) GetUser(writer http.ResponseWriter, request *http.Request) {
 	userId, err := strconv.ParseUint(request.Context().Value("userId").(string), 10, 64)
+
 	if err != nil {
 		http.Error(writer, err.Error(), http.StatusBadRequest)
 		return
 	}
 
-	value, err := controller.userRepository.FindById(userId)
-	if err != nil {
-		if err.Error() == ErrorUserNotFound {
-			http.Error(writer, err.Error(), http.StatusNotFound)
-			return
-		}
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
-	}
+	authUserId, _ := request.Context().Value("auth_userId").(int)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int)
 
-	writer.Header().Set("Content-Type", "application/json")
-	err = json.NewEncoder(writer).Encode(value)
-	if err != nil {
-		http.Error(writer, err.Error(), http.StatusInternalServerError)
+	if ((uint64(authUserId)) == userId) ||
+		(model.Role(authUserRole) == model.Administrator) {
+		value, err := controller.userRepository.FindById(userId)
+		if err != nil {
+			if err.Error() == ErrorUserNotFound {
+				http.Error(writer, err.Error(), http.StatusNotFound)
+				return
+			}
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+		}
+
+		writer.Header().Set("Content-Type", "application/json")
+		err = json.NewEncoder(writer).Encode(value)
+		if err != nil {
+			http.Error(writer, err.Error(), http.StatusInternalServerError)
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
 
@@ -100,36 +109,22 @@ func (controller defaultController) PutUser(writer http.ResponseWriter, request 
 		return
 	}
 
-	if _, err := controller.userRepository.Update(&model.User{
-		Id:    userId,
-		Email: requestData.Email,
-		Name:  requestData.Name,
-	}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-}
+	authUserId, _ := request.Context().Value("auth_userId").(int)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int)
 
-func (controller defaultController) PostUser(writer http.ResponseWriter, request *http.Request) {
-	var requestData JsonFormatCreateUserRequest
-	if err := json.NewDecoder(request.Body).Decode(&requestData); err != nil {
-		writer.WriteHeader(http.StatusBadRequest)
-		return
+	if ((uint64(authUserId)) == userId) ||
+		(model.Role(authUserRole) == model.Administrator) {
+		if _, err := controller.userRepository.Update(&model.User{
+			Id:    userId,
+			Email: requestData.Email,
+			Name:  requestData.Name,
+		}); err != nil {
+			writer.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
-
-	bcryptHasher := crypto.NewBcryptHasher()
-	hashedPassword, _ := bcryptHasher.Hash(requestData.Password)
-
-	if _, err := controller.userRepository.Create(&model.User{
-		Email:    requestData.Email,
-		Password: hashedPassword,
-		Name:     requestData.Name,
-		Role:     model.Customer,
-	}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	writer.WriteHeader(http.StatusCreated)
 }
 
 func (controller defaultController) DeleteUser(writer http.ResponseWriter, request *http.Request) {
@@ -139,9 +134,17 @@ func (controller defaultController) DeleteUser(writer http.ResponseWriter, reque
 		return
 	}
 
-	if err := controller.userRepository.Delete(&model.User{Id: userId}); err != nil {
-		writer.WriteHeader(http.StatusInternalServerError)
-		return
+	authUserId, _ := request.Context().Value("auth_userId").(int)
+	authUserRole, _ := request.Context().Value("auth_userRole").(int)
+
+	if ((uint64(authUserId)) == userId) ||
+		(model.Role(authUserRole) == model.Administrator) {
+		if err := controller.userRepository.Delete(&model.User{Id: userId}); err != nil {
+			writer.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+	} else {
+		writer.WriteHeader(http.StatusUnauthorized)
 	}
 }
